requirements.txt
keras==2.6.0
matplotlib==3.4.2
numpy==1.19.5
pandas==1.3.1
pynrrd==0.4.2
PyYAML==5.4.1
scikit-image==0.18.3
scipy==1.6.2
statsmodels==0.12.2
tensorflow-gpu==2.6.0